requirements.txt
Django==3.2.8
djangorestframework==3.12.4
graphene-django==2.15.0
django-cors-headers==3.10.0
django-heroku==0.3.1
PyMySQL==1.0.2
gunicorn==20.1.0
